Add tests for pod mutator
