+++ pkg/webhook/mutation/pod/metadata/mutator.go
@@ -2,10 +2,8 @@ package metadata
 
 import (
 	"context"
-	"encoding/json"
 	"strings"
 
-	"github.com/Dynatrace/dynatrace-operator/pkg/api/v1beta3/dynakube"
 	"github.com/Dynatrace/dynatrace-operator/pkg/consts"
 	dtingestendpoint "github.com/Dynatrace/dynatrace-operator/pkg/injection/namespace/ingestendpoint"
 	maputils "github.com/Dynatrace/dynatrace-operator/pkg/util/map"
@@ -69,26 +67,9 @@ func (mut *Mutator) Mutate(ctx context.Context, request *dtwebhook.MutationReque
 	setupVolumes(request.Pod)
 	mutateUserContainers(request.BaseRequest)
 	updateInstallContainer(request.InstallContainer, workload, request.DynaKube.Status.KubernetesClusterMEID)
+	propagateMetadataAnnotations(request)
 	setInjectedAnnotation(request.Pod)
 	setWorkloadAnnotations(request.Pod, workload)
-	copyMetadataFromNamespace(request.Pod, request.Namespace, request.DynaKube)
-
-	workloadAnnotations := map[string]string{}
-
-	for key, value := range request.Pod.Annotations {
-		if !strings.HasPrefix(key, dynakube.MetadataPrefix) {
-			continue
-		}
-
-		split := strings.Split(key, dynakube.MetadataPrefix)
-		workloadAnnotations[split[1]] = value
-	}
-
-	workloadAnnotationsJson, _ := json.Marshal(workloadAnnotations)
-	request.InstallContainer.Env = append(request.InstallContainer.Env,
-		corev1.EnvVar{
-			Name: consts.EnrichmentWorkloadAnnotationsEnv, Value: string(workloadAnnotationsJson)},
-	)
 
 	return nil
 }
@@ -150,55 +131,6 @@ func setWorkloadAnnotations(pod *corev1.Pod, workload *workloadInfo) {
 	// https://bitbucket.lab.dynatrace.org/projects/DEUS/repos/semantic-dictionary/browse/source/fields/k8s.yaml
 	pod.Annotations[dtwebhook.AnnotationWorkloadKind] = strings.ToLower(workload.kind)
 	pod.Annotations[dtwebhook.AnnotationWorkloadName] = workload.name
-}
-
-func copyMetadataFromNamespace(pod *corev1.Pod, namespace corev1.Namespace, dk dynakube.DynaKube) {
-	copyMetadataAccordingToCustomRules(pod, namespace, dk)
-	copyMetadataAccordingToPrefix(pod, namespace)
-
-}
-
-func copyMetadataAccordingToPrefix(pod *corev1.Pod, namespace corev1.Namespace) {
-	for key, value := range namespace.Annotations {
-		if strings.HasPrefix(key, dynakube.MetadataPrefix) {
-			setPodAnnotationIfNotExists(pod, key, value)
-		}
-	}
-}
-
-func copyMetadataAccordingToCustomRules(pod *corev1.Pod, namespace corev1.Namespace, dk dynakube.DynaKube) {
-	for _, rule := range dk.Status.MetadataEnrichment.Rules {
-		if rule.Target == "" {
-			log.Info("rule without target set found, ignoring", "source", rule.Source, "type", rule.Type)
-
-			continue
-		}
-
-		var valueFromNamespace string
-
-		var exists bool
-
-		switch rule.Type {
-		case dynakube.EnrichmentLabelRule:
-			valueFromNamespace, exists = namespace.Labels[rule.Source]
-		case dynakube.EnrichmentAnnotationRule:
-			valueFromNamespace, exists = namespace.Annotations[rule.Source]
-		}
-
-		if exists {
-			setPodAnnotationIfNotExists(pod, rule.ToAnnotationKey(), valueFromNamespace)
-		}
-	}
-}
-
-func setPodAnnotationIfNotExists(pod *corev1.Pod, key, value string) {
-	if pod.Annotations == nil {
-		pod.Annotations = make(map[string]string)
-	}
-
-	if _, ok := pod.Annotations[key]; !ok {
-		pod.Annotations[key] = value
-	}
 }
 
 func ContainerIsInjected(container corev1.Container) bool {
